Authorization and authentication using JWT

+++ controllers/api/v1/posts_api.js
@@ -23,19 +23,22 @@ module.exports.destroy = async function(req, res){
   try{
       let post = await Post.findById(req.params.id);
 
-      //if (post.user == req.user.id){  
+      if (post.user == req.user.id){  
           post.remove();  // delete the post
 
           await Comment.deleteMany({post: req.params.id});  // delete it's comments
+
+          
 
           return res.json(200, {
             message: 'Post and associatd comments deleted successfully'
           });
 
-      // }else{
-      //     req.flash('error', 'You cannot delete this post!');
-      //     return res.redirect('back');
-      // }
+      }else{
+          return res.json(401,{
+            message: "You cannot delete this post"
+          });
+      }
 
   }catch(err){
     console.log('********',err);
